upd: remove validation from orm_manager

+++ repository/orm_manager.py
@@ -1,5 +1,4 @@
-from sqlalchemy.orm import sessionmaker, Session
-from sqlalchemy.orm import Query
+from sqlalchemy.orm import sessionmaker
 from sqlalchemy.orm import joinedload
 import logging
 
@@ -11,7 +10,7 @@ from repository.session_context_manager import session_scope
 from group import Group
 from task import Task
 
-from exceptions import (StatusNotFound, GIDNotFound, TIDNotFound, GroupNotFound)
+from exceptions import (GIDNotFound, TIDNotFound, GroupNotFound)
 
 log = logging.getLogger(__name__)
 
@@ -36,8 +35,10 @@ class ORMManager:
 
     def add_task(self, task: Task) -> tuple[int, str, str, str]:
         with session_scope(self.Session, "Add task") as session:
-            group = self._ensure_group_title_exists(session, task.group)
-            task_orm = TaskModel(title=task.title, status=task.status, group = group)
+            found_group = session.query(GroupModel).filter(GroupModel.title == task.group).first()
+            if not found_group:
+                raise GroupNotFound(task.group)
+            task_orm = TaskModel(title=task.title, status=task.status, group = found_group)
             session.add(task_orm)
             session.flush()
             log.info("Add task: SUCCESS; ID=%r, Title=%r, Status=%r, Group=%r", task_orm.id, task_orm.title, task_orm.status, task_orm.group)
@@ -59,11 +60,12 @@ class ORMManager:
             query = session.query(TaskModel)
             if filtered:
                 if status:
-                    status = self._ensure_status_exists(query, status)
                     query = query.filter(TaskModel.status == status)
                 if group:
-                    group = self._ensure_group_title_exists(session, group)
-                    query = query.filter(TaskModel.group_id == group.id)
+                    found_group = session.query(GroupModel).filter(GroupModel.title == group).first()
+                    if not found_group:
+                        raise GroupNotFound(group)
+                    query = query.filter(TaskModel.group_id == found_group.id)
             query = query.order_by(sorting_map[sort_type])
             query = query.options(joinedload(TaskModel.group))
             rows = query.all()
@@ -101,9 +103,11 @@ class ORMManager:
     def delete_task(self, ids: list[int]) -> list[int]:
         with session_scope(self.Session, "Delete task") as session:
             query = session.query(TaskModel)
-            ids = self._ensure_task_id_exists(query, ids)
-            for id in ids:
-                task = query.filter(TaskModel.id == id).first()
+            tasks = query.filter(TaskModel.id.in_(ids)).all()
+            if len(tasks) < len(ids):
+                raise TIDNotFound(ids)
+
+            for task in tasks:
                 session.delete(task)
             log.info("Delete task: SUCCESS; IDs=%r", ids)
             return ids
@@ -112,9 +116,11 @@ class ORMManager:
     def delete_group(self, id: list[int]) -> list[int]:
         with session_scope(self.Session, "Delete group") as session:
             query = session.query(GroupModel)
-            id = self._ensure_group_id_exists(query, id)
-            for id in id:
-                group = query.filter(GroupModel.id == id).first()
+            groups = query.filter(GroupModel.id.in_(id)).all()
+            if len(groups) < len(id):
+                raise GIDNotFound(id)
+            
+            for group in groups:
                 session.delete(group)
             log.info("Delete group: SUCCESS; IDs=%r", id)
             return id
@@ -123,8 +129,9 @@ class ORMManager:
     def set_status(self, ids: list[int], status: str) -> tuple[list[int], str]:
         with session_scope(self.Session, "Set status") as session:
             query = session.query(TaskModel)
-            ids = self._ensure_task_id_exists(query, ids)
             tasks = query.filter(TaskModel.id.in_(ids)).all()
+            if len(tasks) < len(ids):
+                raise TIDNotFound(ids)
             for t in tasks:
                 t.status = status
             log.info("Set status: SUCCESS; IDs=%r, New status=%r", ids, status)
@@ -134,64 +141,26 @@ class ORMManager:
     def format_task(self, ids: list[int], title: str, status: str, group: str) -> tuple[list[int], str, str, str]:
         with session_scope(self.Session, "Format task") as session:
             query = session.query(TaskModel)
-            group = self._ensure_group_title_exists(session, group)
-            ids = self._ensure_task_id_exists(query, ids)
+            found_group = session.query(GroupModel).filter(GroupModel.title == group).first()
+            if not found_group:
+                raise GroupNotFound(group)
             tasks = query.filter(TaskModel.id.in_(ids)).all()
+            if len(tasks) < len(ids):
+                raise TIDNotFound(ids)
             for t in tasks:
                 t.title = title
                 t.status = status
-                t.group = group
-            log.info("Format task: SUCCESS; ID=%r, New title=%r, New status=%r, New group=%r", ids, title, status, group)
-            return ids, title, status, group.title
+                t.group = found_group
+            log.info("Format task: SUCCESS; ID=%r, New title=%r, New status=%r, New group=%r", ids, title, status, found_group.title)
+            return ids, title, status, found_group.title
 
 
     def format_group(self, id: list[int], title: str) -> tuple[list[int], str]:
         with session_scope(self.Session, "Format group") as session:
             query = session.query(GroupModel)
-            id = self._ensure_group_id_exists(query, id)
             group = query.filter(GroupModel.id.in_(id)).first()
+            if len(group) < len(id):
+                raise GIDNotFound(id)
             group.title = title
             log.info("Format group: SUCCESS; ID=%r, New title=%r", id, title)
             return id, group.title
-
-
-    def _ensure_group_title_exists(self, session: Session, title: str) -> GroupModel:
-        query = session.query(GroupModel)
-        group = query.filter(GroupModel.title == title).first()
-        if group:
-            log.debug("Ensure groups title exists: SUCCESS; GroupID=%r, GroupTitle=%r", group.id, group.title)
-            return group
-        else:
-            log.error("Ensure groups title exists: FAILED; GroupTitle=%r", title)
-            raise GroupNotFound(title)
-
-
-    def _ensure_task_id_exists(self, query: Query[TaskModel], ids: list[int]) -> list[int]:
-        found_ids = query.filter(TaskModel.id.in_(ids)).all()
-        if len(found_ids) < len(ids):
-            log.error("Ensure TaskID exists: FAILED; IDs=%r", ids)
-            raise TIDNotFound(ids)
-        else:
-            log.debug("Ensure TaskID exists: SUCCESS; IDs=%r", ids)
-            return ids
-
-
-    def _ensure_group_id_exists(self, query: Query[GroupModel], id: list[int]) -> list[int]:
-        found_ids = query.filter(GroupModel.id.in_(id)).all()
-        if len(found_ids) < len(id):
-            log.error("Ensure GroupID exists: FAILED; IDs=%r", id)
-            raise GIDNotFound(id)
-        else:
-            log.debug("Ensure GroupID exists: SUCCESS; IDs=%r", id)
-            return id
-
-
-    def _ensure_status_exists(self, query: Query[TaskModel], status: str) -> str:
-        exists = query.filter(TaskModel.status == status).first()
-        if exists:
-            log.debug("Ensure status exists: SUCCESS; Status=%r", status)
-            return status
-        else:
-            log.error("Ensure status exists: FAILED; Status=%r", status)
-            raise StatusNotFound(status)
-
